refactor: fix package paths for extractors
