+++ plugins/extractors/grafana/extractor_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 
 	"github.com/odpf/meteor/core/extractor"
-	"github.com/odpf/meteor/logger"
+	"github.com/odpf/meteor/internal/logger"
 	"github.com/odpf/meteor/plugins/extractors/grafana"
 	"github.com/odpf/meteor/proto/odpf/meta"
 	"github.com/stretchr/testify/assert"
