+++ plugins/extractors/kafka/extractor_test.go
@@ -12,7 +12,7 @@ import (
 
 	kafkaLib "github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/odpf/meteor/core/extractor"
-	"github.com/odpf/meteor/logger"
+	"github.com/odpf/meteor/internal/logger"
 	"github.com/odpf/meteor/plugins/extractors/kafka"
 	"github.com/odpf/meteor/plugins/testutils"
 	"github.com/odpf/meteor/proto/odpf/meta"
